+++ graphein/grn/parse_regnetwork.py
@@ -1,12 +1,12 @@
-import functools
-
+"""Functions for parsing GRNs from RegNetwork."""
 # %%
 # Graphein
 # Author: Ramon Vinas, Arian Jamasb <arian@jamasb.io>
 # License: MIT
 # Project Website: https://github.com/a-r-j/graphein
 # Code Repository: https://github.com/a-r-j/graphein
-import logging
+
+import functools
 import os
 import ssl
 import zipfile
@@ -15,21 +15,25 @@ from typing import Callable, List, Optional
 
 import pandas as pd
 import wget
+from loguru import logger as log
 
 from graphein.utils.utils import filter_dataframe, ping
-
-log = logging.getLogger(__name__)
 
 
 def _download_RegNetwork(
     root_dir: Optional[Path] = None, network_type: str = "human"
 ) -> str:
     """
-    Downloads RegNetwork regulatory interactions to the root directory. Returns the filepath.
+    Downloads RegNetwork regulatory interactions to the root directory.
+    Returns the filepath.
 
-    :param root_dir: Path object specifying the location to download RegNetwork to. Default is None which downloads to the dataset/ directory inside graphein.
-    :type root_dir: patlib.Path, optional
-    :param network_type: Specifies whether to download human or mouse regulatory network. Supported values: "human" (default), "mouse".
+    :param root_dir: ``Path`` object specifying the location to download
+        RegNetwork to. Default is ``None`` which downloads to the ``dataset/``
+        directory inside graphein.
+    :type root_dir: pathlib.Path, optional
+    :param network_type: Specifies whether to download human or mouse
+        regulatory network. Supported values: ``"human"`` (default),
+        ``"mouse"``.
     :type network_type: str
     :returns: path to downloaded RegNetwork
     :rtype: str
@@ -39,7 +43,8 @@ def _download_RegNetwork(
     ping_result = ping("regnetworkweb.org")
     if not ping_result:
         log.warning(
-            "RegNetwork is not available. Please check your internet connection or verify at: http://www.regnetworkweb.org"
+            "RegNetwork is not available. Please check your internet \
+            connection or verify at: http://www.regnetworkweb.org"
         )
 
     mouse_url = "https://regnetworkweb.org/download/mouse.zip"
@@ -51,7 +56,8 @@ def _download_RegNetwork(
         url = mouse_url
     else:
         raise ValueError(
-            f"network_type: {network_type} is unsupported. Please use 'human' or 'mouse'"
+            f"network_type: {network_type} is unsupported. Please use 'human' \
+            or 'mouse'"
         )
 
     # If no root dir is provided, use the dataset directory inside graphein.
@@ -81,9 +87,14 @@ def _download_RegNetwork(
 
 def _download_RegNetwork_regtypes(root_dir: Optional[Path] = None) -> str:
     """
-    Downloads RegNetwork regulatory interactions types to the root directory
+    Downloads RegNetwork regulatory interactions types to the root directory.
 
-    :param root_dir: Path object specifying the location to download RegNetwork to
+    :param root_dir: ``Path`` object specifying the location to download
+        RegNetwork to
+    :type root_dir: pathlib.Path, optional. Default is ``None`` which downloads
+        to the ``dataset/`` directory inside graphein.
+    :returns: Path to downloaded RegNetwork
+    :rtype: str
     """
     url = "https://regnetworkweb.org/download/RegulatoryDirections.zip"
 
@@ -117,7 +128,14 @@ def load_RegNetwork_interactions(
     root_dir: Optional[Path] = None,
 ) -> pd.DataFrame:
     """
-    Loads RegNetwork interaction datafile. Downloads the file first if not already present.
+    Loads RegNetwork interaction datafile. Downloads the file first if not
+    already present.
+
+    :param root_dir: ``Path`` object specifying the location of the
+        RegNetwork download.
+    :type root_dir: pathlib.Path, optional. Default is ``None``.
+    :returns: RegNetwork interaction data as a pd.DataFrame.
+    :rtype: pd.DataFrame
     """
     file = _download_RegNetwork(root_dir)
     return pd.read_csv(
@@ -130,7 +148,14 @@ def load_RegNetwork_regulation_types(
     root_dir: Optional[Path] = None,
 ) -> pd.DataFrame:
     """
-    Loads RegNetwork regulation types. Downloads the file first if not already present.
+    Loads RegNetwork regulation types. Downloads the file first if not already
+        present.
+
+    :param root_dir: ``Path`` object specifying the location of the RegNetwork
+        download.
+    :type root_dir: pathlib.Path, optional. Default is ``None``.
+    :returns: RegNetwork regulation types as a pd.DataFrame.
+    :rtype: pd.DataFrame
     """
     file = _download_RegNetwork_regtypes(root_dir)
     return pd.read_csv(
@@ -149,7 +174,12 @@ def parse_RegNetwork(
     Parser for RegNetwork interactions
 
     :param gene_list: List of gene identifiers
-    :return Pandas dataframe with the regulatory interactions between genes in the gene list
+    :type gene_list: List[str]
+    :param root_dir: ``Path`` object specifying the location of the RegNetwork
+        download.
+    :return: Pandas DataFrame with the regulatory interactions between genes
+        in the gene list.
+    :rtype: pd.DataFrame
     """
     # Load dataframes
     df = load_RegNetwork_interactions(root_dir)
@@ -175,11 +205,13 @@ def filter_RegNetwork(
     df: pd.DataFrame, funcs: Optional[List[Callable]] = None
 ) -> pd.DataFrame:
     """
-    Filters results of RegNetwork call by providing a list of user-defined functions that accept a dataframe and return a dataframe
+    Filters results of RegNetwork call by providing a list of user-defined
+    functions that accept a DataFrame and return a DataFrame
 
-    :param df: pd.Dataframe to filter
-    :param funcs: list of functions that carry out dataframe processing
-    :return: processed dataframe
+    :param df: pd.DataFrame to filter
+    :param funcs: List of functions that carry out DataFrame processing.
+    :return: Processed DataFrame.
+    :rtype: pd.DataFrame
     """
     if funcs is not None:
         df = filter_dataframe(df, funcs)
@@ -189,7 +221,7 @@ def filter_RegNetwork(
 
 def standardise_RegNetwork(df: pd.DataFrame) -> pd.DataFrame:
     """
-    Standardises STRING dataframe, e.g. puts everything into a common format
+    Standardises RegNetwork dataframe, e.g. puts everything into a common format
 
     :param df: Source specific Pandas dataframe
     :type df: pd.DataFrame
@@ -217,16 +249,23 @@ def RegNetwork_df(
     filtering_funcs: Optional[List[Callable]] = None,
 ) -> pd.DataFrame:
     """
-    Generates standardised dataframe with RegNetwork protein-protein interactions, filtered according to user's input
-    :return: Standardised dataframe with RegNetwork interactions
+    Generates standardised dataframe with RegNetwork protein-protein
+    interactions, filtered according to user's input.
+
+    :param gene_list: List of gene identifiers.
+    :type gene_list: List[str]
+    :param root_dir: ``Path`` object specifying the location of the RegNetwork
+        dataset.
+    :type root_dir: pathlib.Path, optional. Default is ``None``.
+    :param filtering_funcs: List of (optional) functions that accept a dataframe
+        and return a dataframe. These functions are applied to the dataframe to
+        filter entries according to the user's input.
+    :type filtering_funcs: List[Callable], optional. Default is ``None``.
+    :return: Standardised dataframe with RegNetwork interactions.
+    :rtype: pd.DataFrame
     """
     df = parse_RegNetwork(gene_list=gene_list, root_dir=root_dir)
     df = filter_RegNetwork(df, filtering_funcs)
     df = standardise_RegNetwork(df)
 
     return df
-
-
-if __name__ == "__main__":
-    df = RegNetwork_df(["AATF", "MYC", "USF1", "SP1", "TP53", "DUSP1"])
-    print(df)